made iqmanim framerate a float
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@2383 07af1644-1164-4067-b897-63bf4dc9b748

+++ engine/iqm.h
@@ -89,7 +89,7 @@ struct iqmanim
 {
     uint name;
     uint first_frame, num_frames;
-    uint framerate;
+    float framerate;
     uint flags;
 };
